+++ unit-tests/src/test/scala/scala/scalanative/unsafe/CArrayBoxingTest.scala
@@ -7,7 +7,7 @@ import org.junit.Assert._
 import scalanative.junit.utils.AssertThrows._
 
 import scalanative.unsafe.Nat._
-import scalanative.runtime.toRawPtr
+import scalanative.unsigned._
 import scalanative.libc.stdlib.malloc
 import java.lang.Long.toHexString
 
@@ -15,9 +15,9 @@ class CArrayBoxingTest {
   var any: Any = null
 
   @noinline lazy val nullArr: CArray[Byte, _4] = null
-  @noinline lazy val arr: CArray[Byte, _4] = !malloc(64)
+  @noinline lazy val arr: CArray[Byte, _4] = !malloc(64.toULong)
     .asInstanceOf[Ptr[CArray[Byte, _4]]]
-  @noinline lazy val arr2: CArray[Byte, _4] = !malloc(64)
+  @noinline lazy val arr2: CArray[Byte, _4] = !malloc(64.toULong)
     .asInstanceOf[Ptr[CArray[Byte, _4]]]
 
   @noinline def f[T](x: T): T      = x
