+++ unit-tests/src/test/scala/scala/scalanative/unsafe/ExternTest.scala
@@ -6,6 +6,7 @@ import org.junit.Test
 import org.junit.Assert._
 
 import scalanative.unsafe._
+import scalanative.unsigned._
 
 object ExternTest {
   /* These can be nested inside an object but not a class - see #897
@@ -26,7 +27,7 @@ object ExternTest {
   // workaround for CI
   def runTest(): Unit = {
     import scalanative.libc.string
-    val bufsize = 10L
+    val bufsize = 10.toUInt
     val buf1    = stackalloc[Byte](bufsize)
     val buf2    = stackalloc[Byte](bufsize)
     Ext1.snprintf(buf1, bufsize, c"%s", c"hello")
@@ -44,7 +45,7 @@ class ExternTest {
     val args = Seq("skipped", "skipped", "skipped", "-b", "-f", "farg")
 
     Zone { implicit z =>
-      val argv = stackalloc[CString](args.length)
+      val argv = stackalloc[CString](args.length.toUInt)
 
       for ((arg, i) <- args.zipWithIndex) {
         argv(i) = toCString(arg)
